+++ src/commands.cpp
@@ -2,6 +2,7 @@
 
 #include <iostream>
 
+#include "benchmark.hpp"
 #include "evaluator.hpp"
 #include "iterator.hpp"
 #include "miner.hpp"
@@ -210,13 +211,7 @@ void Commands::iterate(const std::string& count) {
   }
 }
 
-void Commands::collatz(const std::string& path) {
-  initLog(true);
-  Parser parser;
-  Program program = parser.parse(path);
-  Evaluator evaluator(settings);
-  Sequence seq;
-  evaluator.eval(program, seq);
-  bool is_collatz = Miner::isCollatzValuation(seq);
-  std::cout << (is_collatz ? "true" : "false") << std::endl;
+void Commands::benchmark() {
+  Benchmark benchmark;
+  benchmark.all();
 }