+++ src/include/benchmark.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+class Benchmark {
+ public:
+  void all();
+
+  void operations();
+
+  void programs();
+};