+++ CHANGELOG.md
@@ -9,6 +9,8 @@ To install or update LODA, please follow the [installation instructions](https:/
 ### Enhancements
 
 * Internal big number format uses 64-bit words (faster and larger range).
+* Remove hard-coded Collatz checks
+* Enhanced benchmark for comparing releases
 
 ## v21.9.28
 
@@ -26,6 +28,13 @@ To install or update LODA, please follow the [installation instructions](https:/
 
 * Optimized server-side processing of submitted programs.
 * Regularly update OEIS index and programs folder, and regenerate stats. No more restarts of miner processes needed.
+* Benchmark:
+
+| Sequence | Max Terms | Time for 100 terms |
+|----------|-----------|--------------------|
+| A000796  |    283    |      153.1ms       |
+| A001113  |    286    |      133.9ms       |
+| A002193  |    106    |      3566.6ms      |
 
 ## v21.9.21